conncheck: Fix unconnectable streams with early ichecks
When we receive an early incoming check, we store it, then we trigger
it once we receive the remote candidates and credentials. But due to
the retransmit flag being unset in the CandidateCheckPair due to commit
36f306f4 which removed the retransmit init value to TRUE. This caused
the triggerred check to keep the pair in the IN_PROGRESS state without
trigerring the check. This causes a non-agressive nomination agent to
never establish connection because any early icheck derived pairs will
remain in-progress forever.